code/backend/requirements.txt
torch
torchvision
tensorflow
keras
opencv-python
face_recognition
sklearn
Pillow
Flask